The code displayed in the documentation is inconsistent with the example code provided in the repo.  If a user copies and pastes the example from the documentation, then they'll receive either:
```
AttributeError: 'User' object has no attribute 'id'
```
because the attribute name is mismatched,  OR a request will return:

```
{"exception":"InvalidRetrieveUserObject","reasons":"The retrieve_user
method should return either a dict or an object with a to_dict method."}
```
because the User object in the documentation does not have a method to
allow to_dict to be called.

Just a documentation fix for knuckleheads like me, following along at
home.